fix: move hook for languages from core to lib
