--- src/core/flagsmith/hooks/use-target-languages.ts
@@ -1,11 +1,0 @@
-import flagsmith from "flagsmith";
-import FEATURES_IDS_LIST from "../features-ids-list";
-
-const useTargetLanguages = () => {
-	const valueString: string = flagsmith.getValue(FEATURES_IDS_LIST.languages_list);
-	const targetLanguages: string[] = JSON.parse(valueString);
-
-	return targetLanguages;
-};
-
-export default useTargetLanguages;
